Refactoring and remove old tag file
