--- src/main/resources/data/iota/tags/items/bucket_has_mob.json
@@ -1,10 +1,0 @@
-{
-  "replace": false,
-  "values": [
-    "minecraft:pufferfish_bucket",
-    "minecraft:salmon_bucket",
-    "minecraft:cod_bucket",
-    "minecraft:tropical_fish_bucket",
-    "minecraft:axolotl_bucket"
-  ]
-}
